Add missing space after \and with multiple authors, reported by Patrick Massot.
Unfixed but related:
This fixes the problem specifically, but it was also reported that it
took a long time 430s for the error to occurr.
I have yet to reproduce that.  Try with the perfectoid project

Additionally the error reported is fairly obtuse,
which likely has to do with short-circuiting the error and stopping on the first error,
When the information is in some subsequent error.
We should attempt to coax a better one out of tectonic if possible.